feat: integrate Dear ImGui
This commit integrates Dear ImGui into the graphics library, providing an immediate mode GUI for
both Metal and Vulkan backends.

Key changes include:
 - Dear ImGui Integration: Added the necessary backend files for Metal and Vulkan to render ImGui
   draw data. The Device and CommandBuffer interfaces have been extended to support ImGui rendering
   commands.
 - New Example: A new imgui_usage example demonstrates how to initialize and use ImGui within an
   application.
 - CMake Updates:
     - ImGui is now fetched as a dependency using FetchContent.
     - A new GFX_ENABLE_IMGUI option in CMake controls whether to build with ImGui support.
     - The imgui_usage example is only built when GFX_ENABLE_IMGUI is enabled.
 - Backend-Specific Implementations:
     - For Metal, imgui_impl_metal is used to render ImGui.
     - For Vulkan, imgui_impl_vulkan is used.
 - Conditional Compilation: All ImGui-related code is conditionally compiled based on the
   GFX_ENABLE_IMGUI flag to keep the core library clean when ImGui is not needed.

+++ src/Vulkan/VulkanGraphicsPipeline.cpp
@@ -67,7 +67,7 @@ VulkanGraphicsPipeline::VulkanGraphicsPipeline(const VulkanDevice* device, const
         vertexInputBindingDescriptions = {
             vk::VertexInputBindingDescription{}
                 .setBinding(0)
-                .setStride(vertexLayout->stride)
+                .setStride(static_cast<uint32_t>(vertexLayout->stride))
                 .setInputRate(vk::VertexInputRate::eVertex)
         };
         vertexInputAttributeDescriptions.resize(vertexLayout->attributes.size());
@@ -76,7 +76,7 @@ VulkanGraphicsPipeline::VulkanGraphicsPipeline(const VulkanDevice* device, const
                 .setLocation(i)
                 .setBinding(0)
                 .setFormat(toVkFormat(attribute.format))
-                .setOffset(attribute.offset);
+                .setOffset(static_cast<uint32_t>(attribute.offset));
             i++;
         }
         vertexInputStateCreateInfo = vk::PipelineVertexInputStateCreateInfo()
